+++ Tea.Api.Admin/Tea.Api.Service/Collection/CollectionService.cs
@@ -134,70 +134,70 @@ namespace Tea.Api.Service.Collection
         {
             string msg = await _unitOfWork.CollectionRepository.SaveApproveStg(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveApproveSupplier(SaveApproveSupplier _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveApproveSupplier(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveSale(SaveSaleModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveSale(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveSaleRate(SaveSaleRateFixModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveSaleRate(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveSTG(SaveStgModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveSTG(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveStgRate(SaveStgRateFixModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveStgRate(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveStgSale(SaveStgSaleModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveStgSale(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveSupplier(SaveSupplierModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveSupplier(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.SaveSupplierRate(SaveSupplierRateFixModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.SaveSupplierRate(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.UploadSupplierChallan(SaveChallanImageModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.UploadSupplierChallan(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<DataSet> ICollectionService.DateWiseGradeReport(GradeReportModel _input)
@@ -232,14 +232,14 @@ namespace Tea.Api.Service.Collection
         {
             string msg = await _unitOfWork.CollectionRepository.LateralStgSave(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<SaveReturnModel> ICollectionService.VehicleLockSaveMobile(VehicleLockModel _input)
         {
             string msg = await _unitOfWork.CollectionRepository.VehicleLockSaveMobile(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<DataSet> ICollectionService.GetVehicleLockDetails(GetVehicleLockModel _input)
@@ -274,7 +274,7 @@ namespace Tea.Api.Service.Collection
         {
             string msg = await _unitOfWork.CollectionRepository.UpdateTransferStatus(_input);
             string[] msgList = msg.Split(",");
-            return new SaveReturnModel() { Id = Convert.ToInt16(msgList[0]), Message = msgList[1] };
+            return new SaveReturnModel() { Id = Convert.ToInt64(msgList[0]), Message = msgList[1] };
         }
 
         async Task<DataSet> ICollectionService.GetMobileStgData(GetStgHistoryModel _input)
